refactor: 세션 관련 클래스 패키지 이동 개선
