+++ tomcat/src/main/java/com/techcourse/SessionManager.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package com.techcourse;
 
 import com.techcourse.model.User;
 import java.util.HashMap;
